Refactor the folder structure for example applications
