+++ .idea/gradle.xml
@@ -12,6 +12,8 @@
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/buildSrc" />
             <option value="$PROJECT_DIR$/examples" />
+            <option value="$PROJECT_DIR$/examples/kanban" />
+            <option value="$PROJECT_DIR$/examples/kanban/app" />
             <option value="$PROJECT_DIR$/microbenchmark" />
             <option value="$PROJECT_DIR$/yorkie" />
             <option value="$PROJECT_DIR$/yorkie/proto" />
