--- examples/build.gradle.kts
@@ -1,68 +1,0 @@
-plugins {
-    id("com.android.application")
-    kotlin("android")
-}
-
-android {
-    namespace = "dev.yorkie.examples"
-    compileSdk = 33
-
-    defaultConfig {
-        applicationId = "dev.yorkie.examples"
-        minSdk = 23
-        targetSdk = 33
-        versionCode = 1
-        versionName = "1.0.0"
-
-        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-        vectorDrawables {
-            useSupportLibrary = true
-        }
-    }
-    buildTypes {
-        release {
-            isMinifyEnabled = false
-            proguardFiles(
-                getDefaultProguardFile("proguard-android-optimize.txt"),
-                "proguard-rules.pro",
-            )
-        }
-    }
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_11
-        targetCompatibility = JavaVersion.VERSION_11
-    }
-    kotlinOptions {
-        jvmTarget = JavaVersion.VERSION_11.toString()
-    }
-    buildFeatures {
-        compose = true
-        dataBinding = true
-        viewBinding = true
-    }
-    composeOptions {
-        kotlinCompilerExtensionVersion = "1.3.2"
-    }
-    packagingOptions {
-        resources {
-            excludes += "/META-INF/{AL2.0,LGPL2.1}"
-        }
-    }
-}
-
-dependencies {
-    implementation(project(":yorkie"))
-    implementation("androidx.core:core-ktx:1.9.0")
-    implementation("androidx.appcompat:appcompat:1.5.1")
-    implementation("com.google.android.material:material:1.7.0")
-    implementation("androidx.lifecycle:lifecycle-runtime-ktx:2.5.1")
-    implementation("androidx.activity:activity-compose:1.6.1")
-    implementation("androidx.compose.ui:ui:${Versions.composeUi}")
-    implementation("androidx.compose.ui:ui-tooling-preview:${Versions.composeUi}")
-    implementation("androidx.compose.material:material:${Versions.composeUi}")
-    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-android:${Versions.coroutines}")
-    implementation("androidx.lifecycle:lifecycle-viewmodel-ktx:${Versions.viewModel}")
-    implementation("org.jetbrains.kotlinx:kotlinx-collections-immutable:0.3.5")
-    debugImplementation("androidx.compose.ui:ui-tooling:${Versions.composeUi}")
-    debugImplementation("androidx.compose.ui:ui-test-manifest:${Versions.composeUi}")
-}
