+++ examples/kanban/app/src/main/res/values/colors.xml
