+++ examples/kanban/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Fri Dec 16 01:35:54 KST 2022
+distributionBase=GRADLE_USER_HOME
+distributionUrl=https\://services.gradle.org/distributions/gradle-7.4-bin.zip
+distributionPath=wrapper/dists
+zipStorePath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
